Break NS short channels and admin channels migration apart
